+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Connect SDK Android Changelog
 
-## 1.3.1 - 14 July 2014
+## 1.3.2 -- 6 Aug 2014
+
+- Added launchYouTube(String contentId, float startTime, AppLaunchListener listener) method to Launcher capability
+- Decoupled Netcast and DLNA services
+- Miscellaneous bug fixes
+- [See commits between 1.3.1 and 1.3.2](https://github.com/ConnectSDK/Connect-SDK-Android/compare/1.3.1...1.3.2)
+
+[View files at version 1.3.2](https://github.com/ConnectSDK/Connect-SDK-Android/tree/1.3.2)
+
+## 1.3.1 -- 14 July 2014
 
 - Significant performance fixes with regards to Zeroconf discovery
 - Miscellaneous bug fixes
@@ -8,7 +17,7 @@
 
 [View files at version 1.3.1](https://github.com/ConnectSDK/Connect-SDK-Android/tree/1.3.1)
 
-## 1.3.0 - 23 June 2014
+## 1.3.0 -- 23 June 2014
 
 - Added support for Apple TV
  + Supports media playback & control via HTTP requests
@@ -23,7 +32,7 @@
 
 [View files at version 1.3.0](https://github.com/ConnectSDK/Connect-SDK-Android/tree/1.3.0)
 
-## 1.2.1 - 14 May 2014
+## 1.2.1 -- 14 May 2014
 
 - Fixed numerous issues with Connectable Device Store
 - Added ability to probe for app presence on Roku & DIAL